requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
scipy==1.11.3
mlxtend==0.22.0
openpyxl==3.1.2